pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "zmq_logger"
version = "0.0.1"
authors = [
  { name="Phil Birch", email="dr.phil.birch@gmail.com" },
]
dependencies = ["loguru", "pyzmq","securepickle"]
description = "A small network logger and client that uses loguru and zmq"
readme = "README.md"
requires-python = ">=3.7"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

[project.urls]
"Homepage" = "https://github.com/docPhil99/zmq_logger"
"Bug Tracker" = "https://github.com/docPhil99/zmq_logger/issues"
